requirements.txt
playwright>=1.40.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
flask>=3.0.0
werkzeug>=3.0.0
python-telegram-bot>=20.0
pyinstaller>=6.0.0

